+++ include/cry/utils.h
@@ -14,12 +14,12 @@
  *
  * If input length is n blocks then output length is n+1 blocks.
  *
- * @out     Cipher-text output.
- * @in      Plain-text input.
- * @len     Length of the plain-text (in).
- * @key     Key-Encryption Key (KEK).
- * @keylen  KEK length.
- * @return  0 on success, -1 on failure
+ * @param out    Cipher-text output.
+ * @param in     Plain-text input.
+ * @param len    Length of the plain-text (in).
+ * @param key    Key-Encryption Key (KEK).
+ * @param keylen KEK length.
+ * @return       0 on success, -1 on failure
  */
 int cry_aes_wrap(unsigned char *out, const unsigned char *in, size_t len,
                  const unsigned char *key, size_t keylen);
@@ -29,12 +29,12 @@ int cry_aes_wrap(unsigned char *out, const unsigned char *in, size_t len,
  *
  * If input length is n blocks then output length is n+1 blocks.
  *
- * @out     Plain-text output.
- * @in      Cipher-text input.
- * @len     Length of the plain-text (out).
- * @key     Key-Encryption Key (KEK).
- * @keylen  KEK length.
- * @return  0 on success, -1 on failure
+ * @param out    Plain-text output.
+ * @param in     Cipher-text input.
+ * @param len    Length of the plain-text (out).
+ * @param key    Key-Encryption Key (KEK).
+ * @param keylen KEK length.
+ * @return       0 on success, -1 on failure
  */
 int cry_aes_unwrap(unsigned char *out, const unsigned char *in, size_t len,
                    const unsigned char *key, size_t keylen);