+++ CHANGELOG.md
@@ -21,7 +21,21 @@ Given a version number MAJOR.MINOR.PATCH
 * MINOR incremented for new functionalities
 * PATCH incremented for bug fixes
 
-[0.1.0] - Unreleased
+
+[0.1.1] - Unreleased
+--------------------
+
+Added
+
+* ARC4 stream cipher
+* Elgamal digital signature
+
+Fixed
+
+* MPI "is_prime" function was not working for primes < 1010 (decimal)
+
+
+[0.1.0] - 2019-09-23
 --------------------
 
 Added